Crate kafka_file_writer

Kafka file writer

Listen to Kafka, write the data into NeXus files.

The data may include things like event-mode data, histogrammed views of the event-mode data, sample-environment block data, run metadata, static datasets describing instrument or sample state.

Architecture overview

The architecture of this filewriter is loosely based on the ESS filewriter. The following design elements are architecturally similar to the ESS filewriter:

Pooled architecture driven by a Kafka topic

Configurable nexus_template, sent as part of the run start message

Writer-modules which describe how to write Kafka messages to NeXus files

However, there are also a number of areas where we diverge from the ESS design:

Our writer modules may receive multiple schemas, perhaps from multiple topics. For example, an event-data writer may need ev44, pu00, and vc00 messages as inputs, from _events and _vetoConfig topics.

Our messages may be delivered to multiple writer-modules (for example, a histogramming module and also an event-writing module).

Support for intermediate files (autosave files & explicit store operations).

File-writer

The file-writer is run using:

kafka_file_writer --config config.toml

Or for a development build:

cargo run --bin kafka_file_writer -- --config config.toml

Configuration

Global configuration (e.g. Kafka broker settings) are specified in config.toml.

Per-run configuration, for example NeXus structure, comes via the run start message.

Nexus structures

The file-structure written by this filewriter is defined dynamically, by the nexus_structure field of a run start message.

For example, the nexus_structure defines:

Whether event-mode data is written, and if so, where in the file it is written to

Whether histogram-mode data is written

Which blocks are written

Which static datasets are written

The NeXus structures expected by this program are inspired by, but differ from, the ESS’ NeXus structures.

Some example structures are available in the structures/ directory.

A standalone structure_verify executable is available, which attempts to instantiate an in-memory file with the specified structure, to verify it instantiates correctly. Run it with:

structure_verify --file structures/isis.json

Or for a development build:

cargo run --bin structure_verify -- --file structures/isis.json

Threading model

This filewriter is single-threaded, but is scalable via different filewriter processes writing different runs (potentially on different servers) at the same time.

The hdf5 library, even when built in a threadsafe configuration, can only write to one dataset at a time (threadsafe builds guard all library calls with a global lock). Therefore, there is no significant benefit to HDF5 write speeds when multithreading.

Writer modules may use multithreading internally for parallelizable work.

Modules

config

Global file-writer configuration file support.

error

Top-level error definitions for errors encountered while processing a run.

file_writer

Top-level file writer.

hdf

HDF5-related utilities.

run_writer

Contains the top-level logic for writing a run.

stream

Contains stream consumers.

subscription

Subscription to a Kafka stream by a writer-module.

writer_module

Traits and types used by writer modules.

writer_module_factories

Registry of writer modules usable by the filewriter at runtime.

writer_modules

Contains the definitions of all writer modules implemented by this file-writer.