kafka_file_writer::config

Struct GlobalConfig

pub struct GlobalConfig {
    pub stream_seek_back_in_time_ms: u64,
    pub job_pool_topic: String,
    pub job_pool_consumer_poll_time_ms: u64,
    pub data_consumer_poll_time_ms: u64,
    pub stream_error_backoff_ms: u64,
    pub kafka_assignment_timeout_ms: u64,
    pub job_pool_kafka_consumer_settings: HashMap<String, String>,
    pub data_kafka_consumer_settings: HashMap<String, String>,
    pub file_output_directory: PathBuf,
    pub forced_nexus_structure_filepath: Option<PathBuf>,
    pub exit_after_writing_one_file: bool,
}

Global configuration parameters which control file-writer behavior for all runs.

Fields
stream_seek_back_in_time_ms: u64

How far to seek-back and replay all streams before run start. This is applied globally to all topics.

job_pool_topic: String

Kafka topic to listen for jobs on

job_pool_consumer_poll_time_ms: u64

Kafka poll() timeout for job pool.

data_consumer_poll_time_ms: u64

Kafka poll() timeout for data consumer.

stream_error_backoff_ms: u64

How long to back off for if we got a stream error from Kafka

kafka_assignment_timeout_ms: u64job_pool_kafka_consumer_settings: HashMap<String, String>

Kafka consumer settings for job pool

data_kafka_consumer_settings: HashMap<String, String>

Kafka consumer settings for data consumer

file_output_directory: PathBuf

Directory into which to write NeXus files

forced_nexus_structure_filepath: Option<PathBuf>

Override the nexus_structure in the run start message with the structure specified in a local file

exit_after_writing_one_file: bool
Implementations

impl GlobalConfig

pub fn from_path(path: impl AsRef<Path>) -> Result<Self, ConfigError>

Load a configuration from TOML at the specified filepath.