kafka_file_writer

Module run_writer

Contains the top-level logic for writing a run.

Modules

in_progress_file

Representation of a file which is in the process of being written to.

job_pool

Interface to the job-pool, for acquiring a run to write.

message_router

Message router: routes messages by topic and key to writer modules that declared those as subscriptions.

nexus_structure

Support for parsing a nexus_structure from a run start message.

placed_writer

Representation of a Writer module with a specific NeXus location it is writing to.

run_start_parameters

Parameters from the start of the run currently being written.

single_file_write_task

The full write-task lifecycle for a single run.