kafka_file_writer::run_writer

Module message_router

Message router: routes messages by topic and key to writer modules that declared those as subscriptions.

Naively, we could simply pass every message to every writer module, and writer modules would be responsible for ignoring irrelevant messages. However, the naive approach is not performant for the case where we have many IXseblock writers and corresponding f144 messages (it would scale as N^2)

Instead, allow writer modules to subscribe to topics+keys specified in advance. Use a hashmap lookup to determine the appropriate writer(s) for this key. Writers may subscribe to all keys if they need to receive all messages from a topic.

Structs

MessageRouter

Message router for all topics.

Type Aliases

WriterId

Unique identifier for a specific writer module (which might, for example, be an index into some external Vec).