kafka_file_writer::writer_module

Trait WriterModule

pub trait WriterModule: Debug {
    // Required methods
    fn on_message(
        &mut self,
        meta: &KafkaMessageMeta<'_>,
        message: &DeserializedMessage<'_>,
    ) -> Result<(), WriterModuleError>;
    fn on_run_stop(
        &mut self,
        meta: &KafkaMessageMeta<'_>,
        message: &RunStop<'_>,
    ) -> Result<(), WriterModuleError>;
    fn flush(&mut self) -> Result<(), WriterModuleError>;
    fn finished(&self) -> bool;
}

Runtime behaviour definitions for this writer module type.

Required Methods

fn on_message( &mut self, meta: &KafkaMessageMeta<'_>, message: &DeserializedMessage<'_>, ) -> Result<(), WriterModuleError>

Called for each received message which matched a Subscription provided by this module.

The implementation of this method may write directly to the NeXus file, or buffer data in-memory.

fn on_run_stop( &mut self, meta: &KafkaMessageMeta<'_>, message: &RunStop<'_>, ) -> Result<(), WriterModuleError>

Called when a run-stop message is received.

The implementation of this method may write directly to the NeXus file, or buffer data in-memory.

fn flush(&mut self) -> Result<(), WriterModuleError>

Ensure that any local state is reflected in the NeXus file, in an internally-consistent way.

For example, this may write an in-memory histogram to the file, or write in-memory summary statistics such as min/max/average.

After this method completes, the file must be in a self-consistent state and as up-to-date as possible with data received so far.

This may be implemented as a no-op if this writer module always streams all data to file directly and does not buffer values in-memory.

This may be called multiple times during a single run, for example to generate an autosave/intermediate file and then after run-end.

fn finished(&self) -> bool

Has this writer module finished writing, after a run-stop was received?

Due to out-of-order Kafka message delivery, a run stop may be received before messages which occured during the run. This hook allows a writer module to declare whether it has received all necessary data up to run end. It should return true if all necessary data has been received, or false otherwise.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".