kafka_file_writer::writer_module

Trait WriterModuleSpec

pub trait WriterModuleSpec: WriterModule + Sized {
    type Config: DeserializeOwned;

    const NAME: &'static str;

    // Required method
    fn create(
        parent_group: Group,
        run_start_parameters: &RunStartParameters,
        config: Self::Config,
    ) -> Result<WriterModuleCreateResult<Self>, WriterModuleCreationError>;
}

Creation parameters for writer modules of this type. This trait should be implemented by all writer modules, and is necessary in order to add the writer module to the WriterModuleRegistry.

Required Associated Constants

const NAME: &'static str

The name of this writer module, as referred to in nexus_structure. This must be unique among all writer modules.

Required Associated Types

type Config: DeserializeOwned

The configuration parameters for this type. This type will be automatically deserialized from the JSON parameters specified in nexus_structure.

Required Methods

fn create( parent_group: Group, run_start_parameters: &RunStartParameters, config: Self::Config, ) -> Result<WriterModuleCreateResult<Self>, WriterModuleCreationError>

Create the writer module.

This should return Err(WriterModuleFactoryError::ConfigInvalid) if the configuration is semantically invalid or inconsistent.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".